TopSortInterface::setCircularInterceptor - Use PHP 5.3 typing
The use of keyword `callable` breaks in PHP 5.3.

The type information is still conveyed by the the docblock (`@param`) but
can't be enforced in 5.3

This is the only tricky change for 5.3 compatibility.  Formally, it's a
compatibility break because anyone that implements `TopSortInterface` would
need to adjust their definition of `setCircularInterceptor()`.  In practice,
I'm skeptical that there'd much of an audience that implements
`TopSortInterface` outside of `topsort`.  (Personally, I find `topsort`
appealing because it provides a library of implementations.)